requirement.txt
beautifulsoup4==4.10.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
jieba==0.42.1
mongoengine==0.23.1
numpy==1.21.5
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
torch==1.10.1
transformers==4.15.0
dnspython==2.1.0
